pyproject.toml
[tool.poetry]
name = "eda"
version = "0.1.0"
description = "Exploratory Data Analysis CLI tool"
authors = ["Jes Hunsballe <jes.hunsballe@gmail.com>"]
readme = "README.md"
packages = [{include = "eda", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
pandas = "^2.0.0"
gspread = "^5.12.0"
pandas-gbq = "^0.19.2"
google-auth = "^2.23.0"
google-auth-oauthlib = "^1.0.0"
pyfiglet = "^1.0.2"
ollama = "^0.1.4"
plotly = "^5.18.0"
pyyaml = "^6.0.1"
rich-click = "^1.8.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"

[tool.poetry.scripts]
eda = "eda.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
dependencies = [
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "gspread>=5.0.0",
]
